Change how npm package is referenced
The referenced npm package is automatically resolved by the npm and therefore it's more future proof.

For example, if you install using npm 3 and flat hierarchy, you might encounter a problem if the package is moved to an upper level. This would happen if you have a project that is using an npm module that has a FontFaceObserver dependency.