Fixed ruler for RTL pages
On pages with right-to-left direction on the HTML element, creating an element "out-of-view" by absolutely positioning it to left:-999px will create scrollbars (while the element exists in the DOM). This can be visible to the user and also mess up measurements of the body's scrollWidth that happen in other code.