Do not hide rulers by positioning them on the left of the viewport.
A ruler that is positioned offscreen with left: -999px causes
scrollbars on the page when the page is in right-to-left mode
(i.e. dir="rtl"). Instead of figuring out the directionality
of the page we just position the rulers on top of the viewport.
This shouldn't cause any additional problems because we already
have an existing top: -999px in place.

Closes #103.